refactor: make decomposition
